+++ templates/news/news_add.php
@@ -243,16 +243,6 @@ if ($news_mode == 'edit') {
         </div><!-- /.row -->
     </form>
 
-    <?php
-    if (!defined('BS_VERSION') || BS_VERSION <= 3) { $snSuffix = ''; }
-    elseif (BS_VERSION == 4)                        { $snSuffix = '-bs4'; }
-    else                                            { $snSuffix = '-bs5'; }
-    ?>
-    <link href="/css/summernote<?php echo $snSuffix; ?>.min.css" rel="stylesheet">
-    <script src="/js/summernote/summernote<?php echo $snSuffix; ?>.min.js" defer></script>
-    <script src="/js/summernote/lang/summernote-de-DE.min.js" defer></script>
-    <script src="/js/summernote-init.js" defer></script>
-
     <?php endif; ?>
 
 </div><!-- /.container -->